Refactored module: - fix error - removing a duplicate code - decomposition of functionality - structuring of state and hooks
